chore: Move spline and paintbrush interfaces to toolboxes
