+++ src/toolboxes/paintbrush/index.ts
@@ -1,3 +1,4 @@
 export { PaintbrushCanvas } from "./Canvas";
 export { PaintbrushUI } from "./UI";
 export { Toolbar as PaintbrushToolbar } from "./Toolbar";
+export { BrushStroke } from "./interfaces";
