+++ src/toolboxes/spline/index.ts
@@ -1,3 +1,4 @@
 export { SplineCanvas, events } from "./Canvas";
 export { SplineUI } from "./UI";
 export { Toolbar as SplineToolbar } from "./Toolbar";
+export { Spline } from "./interfaces";
